+++ src/ContentUtils.ts
@@ -10,7 +10,7 @@ import * as fs from "fs";
 export const CONTENT_FOLDER = path.join("webview", "out");
 
 export async function getPreviewTemplate(context: ExtensionContext, templateName: string): Promise<string> {
-    let previewPath = context.asAbsolutePath(path.join(CONTENT_FOLDER, templateName));
+    const previewPath = context.asAbsolutePath(path.join(CONTENT_FOLDER, templateName));
 
     return new Promise<string>((resolve, reject) => {
         fs.readFile(previewPath, "utf8", function (err, data) {