+++ src/test/runTest.ts
@@ -1,8 +1,9 @@
 import * as path from 'path';
 
 import { runTests } from 'vscode-test';
+import { URI } from 'vscode-uri';
 
-async function main() {
+async function main(): Promise<void> {
 	try {
 		// The folder containing the Extension Manifest package.json
 		// Passed to `--extensionDevelopmentPath`
@@ -12,8 +13,11 @@ async function main() {
 		// Passed to --extensionTestsPath
 		const extensionTestsPath = path.resolve(__dirname, './suite/index');
 
-		const launchArgs = process.argv.slice(2);
-		console.log(`Passing command line arguments: ${launchArgs}`);
+		// The path to the workspace, where the files will be created
+		const folderUri = "--folder-uri=" + URI.file(path.resolve(extensionDevelopmentPath, 'src/test/workspace'));
+		const launchArgs = [folderUri, "--disable-extensions"];
+
+		console.log(`Passing command line arguments: --extensionDevelopmentPath=${extensionDevelopmentPath} --extensionTestsPath=${extensionTestsPath} ${launchArgs.join(' ')}`);
 
 		// Download VS Code, unzip it and run the integration test
 		await runTests({ extensionDevelopmentPath, extensionTestsPath, launchArgs });